fix: 뷰대응 수정 및 DiaryReplyView를 DiaryCheckView 뷰에서 분리
